Move telemetry to sub-namespace
Move telemetry-related things to a sub-namespace.
