+++ src/LondonTravel.Site/Telemetry/SiteTelemetryInitializer.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Martin Costello, 2017. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-namespace MartinCostello.LondonTravel.Site
+namespace MartinCostello.LondonTravel.Site.Telemetry
 {
     using Extensions;
     using Microsoft.ApplicationInsights.Channel;
